refactor: etna.analysis.decomposition
