+++ tests/test_analysis/test_decomposition/test_plots.py
@@ -3,6 +3,7 @@ from sklearn.linear_model import LinearRegression
 from sklearn.linear_model import TheilSenRegressor
 
 from etna.analysis import plot_trend
+from etna.analysis import seasonal_plot
 from etna.transforms import ChangePointsTrendTransform
 from etna.transforms import LinearTrendTransform
 from etna.transforms import STLTransform
@@ -30,3 +31,19 @@ def test_plot_bin_seg(example_tsdf, detrend_model):
 @pytest.mark.parametrize("period", (7, 30))
 def test_plot_stl(example_tsdf, period):
     plot_trend(ts=example_tsdf, trend_transform=STLTransform(in_column="target", period=period))
+
+
+@pytest.mark.parametrize(
+    "freq, cycle, additional_params",
+    [
+        ("D", 5, dict(alignment="first")),
+        ("D", 5, dict(alignment="last")),
+        ("D", "week", {}),
+        ("D", "month", {}),
+        ("D", "year", {}),
+        ("M", "year", dict(aggregation="sum")),
+        ("M", "year", dict(aggregation="mean")),
+    ],
+)
+def test_dummy_seasonal_plot(freq, cycle, additional_params, ts_with_different_series_length):
+    seasonal_plot(ts=ts_with_different_series_length, freq=freq, cycle=cycle, **additional_params)
